Refactor collection mappers
