+++ src/AutoMapper.Collection/Collection/ICollectionMapper.cs
@@ -2,21 +2,18 @@ using System;
 using System.Collections.Generic;
 using System.Linq.Expressions;
 
-namespace AutoMapper.EquivalencyExpression
+namespace AutoMapper.Collection
 {
-    public interface IEquivalentExpression
+    public interface ICollectionMapper
     {
     }
 
-    public interface IEquivalentExpression<TSource, TDestination> : IEquivalentExpression
+    public interface ICollectionMapper<TSource, TDestination> : ICollectionMapper
     {
-        bool IsEquivalent(TSource source, TDestination destination);
+        Expression<Func<TSource, TDestination, bool>> EquivalentExpression { get; }
 
         TDestinationItem Map<TSourceItem, TDestinationItem>(TSourceItem source, TDestinationItem destination, ResolutionContext context)
             where TSourceItem : IEnumerable<TSource>
             where TDestinationItem : class, ICollection<TDestination>;
-
-        Expression<Func<TDestination, bool>> ToSingleSourceExpression(TSource destination);
-
     }
 }
