+++ src/AutoMapper.Collection/Collection/Internal/Extensions/TypeExtensions.cs
