+++ src/AutoMapper.Collection/Mappers/EquivalentExpressionAddRemoveCollectionMapper.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 using AutoMapper.Collection;
+using AutoMapper.Collection.Internal.Extensions;
 using AutoMapper.EquivalencyExpression;
 using static System.Linq.Expressions.Expression;
 
@@ -11,28 +12,23 @@ namespace AutoMapper.Mappers
 {
     public class EquivalentExpressionAddRemoveCollectionMapper : IConfigurationObjectMapper
     {
-        private readonly CollectionMapper _collectionMapper = new CollectionMapper();
-
-        public IConfigurationProvider ConfigurationProvider { get; set; }
-
-        public static TDestination Map<TSource, TSourceItem, TDestination, TDestinationItem>(TSource source, TDestination destination, ResolutionContext context, IEquivalentExpression<TSourceItem, TDestinationItem> equivalencyExpression)
-            where TSource : IEnumerable<TSourceItem>
-            where TDestination : class, ICollection<TDestinationItem>
-        {
-            return equivalencyExpression.Map(source, destination, context);
-        }
-
         private static readonly MethodInfo _mapMethodInfo = typeof(EquivalentExpressionAddRemoveCollectionMapper).GetRuntimeMethods().First(_ => _.IsStatic);
+        private readonly CollectionMapper _collectionMapper = new CollectionMapper();
+        public IConfigurationProvider ConfigurationProvider { get; set; }
 
         public bool IsMatch(TypePair typePair)
         {
             return typePair.SourceType.IsEnumerableType()
                    && typePair.DestinationType.IsCollectionType()
-                   && this.GetEquivalentExpression(TypeHelper.GetElementType(typePair.SourceType), TypeHelper.GetElementType(typePair.DestinationType)) != null;
+                   && this.GetCollectionMapper(TypeHelper.GetElementType(typePair.SourceType), TypeHelper.GetElementType(typePair.DestinationType)) != null;
         }
-        
-        public Expression MapExpression(IConfigurationProvider configurationProvider, ProfileMap profileMap, PropertyMap propertyMap,
-            Expression sourceExpression, Expression destExpression, Expression contextExpression)
+
+        public Expression MapExpression(IConfigurationProvider configurationProvider,
+            ProfileMap profileMap,
+            PropertyMap propertyMap,
+            Expression sourceExpression,
+            Expression destExpression,
+            Expression contextExpression)
         {
             var sourceType = TypeHelper.GetElementType(sourceExpression.Type);
             var destinationType = TypeHelper.GetElementType(destExpression.Type);
@@ -48,15 +44,22 @@ namespace AutoMapper.Mappers
                 throw new ArgumentException($"Destination type '{destinationType.FullName}' is not mapped, use type '{typeMap.DestinationType.FullName}' instead.");
             }
 
-            var equivalencyExpression = this.GetEquivalentExpression(typeMap);
+            var collectionMapper = this.GetCollectionMapper(typeMap);
 
             var map = Call(null,
                 _mapMethodInfo.MakeGenericMethod(sourceExpression.Type, sourceType, destExpression.Type, destinationType),
-                    sourceExpression, destExpression, contextExpression, Constant(equivalencyExpression));
+                sourceExpression, destExpression, contextExpression, Constant(collectionMapper));
             var collectionMap = _collectionMapper.MapExpression(configurationProvider, profileMap, propertyMap, sourceExpression, destExpression, contextExpression);
 
             var notNull = NotEqual(destExpression, Constant(null));
             return Condition(notNull, map, collectionMap);
+        }
+
+        public static TDestination Map<TSource, TSourceItem, TDestination, TDestinationItem>(TSource source, TDestination destination, ResolutionContext context, ICollectionMapper<TSourceItem, TDestinationItem> equivalencyExpression)
+            where TSource : IEnumerable<TSourceItem>
+            where TDestination : class, ICollection<TDestinationItem>
+        {
+            return equivalencyExpression.Map(source, destination, context);
         }
     }
 }
